Reduce lifetime parameter virality by splitting Command enum.
Move two variants that take a &[u8] slice into a separate enum
BufCommand, so that Command no longer needs a lifetime parameter. This
should allow code that don't use the buffer commands to not be infected
with lifetime parameters.

Also move the ok_command! macro to mod level for cleanliness.