Implement Config, Display, and Region.
Config provides a builder to create a struct that is then used to send a
series of initialization or configuration commands to the display,
typically only at reset time.

Display encapsulates the display interface and provides a method for
configuring it using `Config`, as well as for creating `Region`s to do
drawing in.

Region represents a rectangular region of the display into which pixel
data is drawn.