Add support for display offset.
This allows the `Display` interface's idea of the display's origin to be
positioned correctly when the display module wires the physical rows and
columns of the display onto the row and com driver lines in an offset
fashion.

The row offset is compensated by `SetDisplayOffset` command which is
built-in to the hardware. Column offset is handled in software.