Add overscanned region support.
An overscanned region will accept pixel data as though it has any
rectangular position and size, but will transparently crop off pixels
that lie outside of the viewable area of the display.

This obviates the need for worrying about boundary conditions if regions
are dynamically positioned in such a way that they could fall partially
or completely outside of the displayable range.