Migrate to v2 fallible GPIO traits, and stop punting on error reporting.
In particular, properly-typed errors are reported from all API surfaces rather
than the crappy practice of squashing everything to `Err(())` that was
happening before. This includes reporting logic errors like out-of-range
arguments to display commands, as well as preserving and propagating errors
from `DisplayInterface`, which in turn propagates them from the underlying SPI
and GPIO interfaces.